ERROR related  to the CSCS submission fixed
k-evolution on CSCS has issue with

if (strncmp(qname,"fld",strlen("fld")) == 0 && qname != NULL)
{
}
else
{
}
While it works perfectly on my computer!
So I removed that condition!

Everything is checked and now we get completely consistent results for the Euclid runs!

Todo:
however there is still the problem of nu + DE which is from the radiation.cpp file! Need to be looked at! There is not impact in our Euclid runs though!