the loop for nKe_numsteps is added and tested
Todo:

- [ ] Compare the phi and phi’ transfer functions with hiclass! [The solutions should be the same at high redshifts otherwise there is a problem!]
- [ ] Why mathematica cannot solve the ODE? Can we fix the mathematica problem?
- [ ] Check the avg of X in kev and hiclass!
- [ ] Also check why mathematics solution and k-volution do not agree?!
- [ ] Fix TK_phi and Tk_phi’ from k-evolution at high redshift! The should be the same!
- [ ] Provide the IC for   TK_phi and Tk_phi’  from hiclass! While it’s interesting considering small initial values for the blowup tests, it’s still important to be provided by hicalss at the first place!

_____________________________

- [x] Make sure that phi’ from k-evolution and phi-phi_old/dtau agree!
They agree now! after fixing nKe_numsteps problem!

- [x] Turning on stochasticity to see whether the avg solution still remains the same as hiclass?
Works!

- [x] Power spectra for the fields from k-evolution/hiclass:
 tension:

While the solution might be correct, the powers are not the same! so we have to provide the IC from hiclass for this test!

It’s because we consider 0 initial conditions for phi_smg and small for phi_’
- [x] What about the speed of sound and w_0? do the k-evolution results agree with hiclass?
A tension:
This might be due to the unit problem! I’m pretty sure about the c_s^2, I think something is considered in hiclass like some unit adjustments in the formula?
We tested with an algebraic equation that we we get from k-evolution makes sense and it’s what we should get! Something needs to be added?