Adding comments about the backward and forward solutions by Emilio
the very simple case of  the matter evolution in a Einstein de Sitter universe.
We know that the solutions should be
\delta\propto a
and
\delta\propto a^(-2/3)

(note that the time variable is n=ln(a) to increase the stability of the solution)

In each of these plots we have 4 curves, labelled as: "G" or "D" for growing or decaying mode respectively and "forward" or "backward" to specify the time direction of the integration.

Let me comment these plots:
normal.pdf: this show the expected behaviour. We select the initial conditions to be either pure growing or pure decaying modes. Both forward and backward integration give the same result. (remember that this is a very simple differential equation)
contaminated.pdf: this is the same plot as before but with a little contamination (10^-6) of the other mode. We can see that at some point the backward integration of the growing mode curve picks the decaying mode solution (which is the growing one back in time) and starts following this
early_times.pdf: this is as the normal case but anticipating the initial time. Here we are plotting the absolute value of \[Delta].  You can see that there is a sign change in \[Delta] for a couple of curves. In particular, they are the curves that decrease in the time variable (the decaying mode with forward evolution, and the growing mode with backward evolution). This is because, as they approach zero, these curves cross it by numerical error. After crossing it they start following the corresponding growing mode.