Todo and updated settings
Todo:
For large boxsize in k-evolution and most probably in gevolution as well, the time-step get large and as a result we get bad behaviour around z~0 because there is a call from CLASS/hiclass at negative redshifts. This should be fixed by considering extrapolation at z~0 instead of interpolation which takes average of one negative redshift.